- Added benchmark applications - Added first attempt at bicubic interpolation   * Continuous discrete convolution kernels - Added the beginnings of Laplacian feature detectors

+++ ip/src/gecv/struct/image/ImageSInt32.java
@@ -58,7 +58,7 @@ public class ImageSInt32 extends ImageInteger<ImageSInt32> {
 		if (!isInBounds(x, y))
 			throw new ImageAccessException("Requested pixel is out of bounds");
 
-		data[getIndex(x, y)] = (short) value;
+		data[getIndex(x, y)] = value;
 	}
 
 	@Override
